Add user auth skeleton
This commit implements the boilerplate for implementing a custom
user model as described in the django documentation for custom
user authentication. It is recommended that these elements be
added to new projects even if the default User model is acceptable.

Oh, and this commit also updates the static root in the production profile.