Migrate to proxy model for trooporg user relation.
As no data is currently or is planned to be stored under the
ScoutMembership model, it can be readily substituted with a proxy
model to the user model. The migrations add in this commit ensure
that all relationships to the user model through a ScoutMembership
are safely converted into direct keys to the user model.

Performing this migration removes the need to perform an additional
database query for User info in several cases. It may, therefore,
provide a slight performance benefit.