Update capitalization of the main site header
For some queries, google uses the first header that it finds on the page
as the title for the search result. This has resulted in the troop's
website being displayed with a rather unprofessional looking all-
lowercase title in search result. This commit adds explicit
capitalization to the main site header to mitigate this issue.