Add explicit 'self' directives to CSP script-src and style-src
Previously, the content-security-policy header specific no script-src
or style-src directives, and so it inherited from default-src to
determine what javascript and css sources were allowed. When configuring
the CKeditor for use on the flatpage admin site, 'unsafe-inline' rules
were appended to the script-src and style-src directives. This caused
there directives to be emitted with only the 'unsafe-inline' values,
and therefore the default-src values were no longer inherited. This
issue was only discovered once the production settings were used since
the development settings already add a 'self' value to script-src for
the django debug toolbar.

This commit adds explicit values to the srcipt-src and style-src
CSP directives to ensure that no additional issues occur in the future
when locally overriding CSP directives.